requirements.txt
asgiref==3.3.1
certifi==2020.12.5
chardet==4.0.0
cos-python-sdk-v5==1.9.4
crcmod==1.7
defusedxml==0.6.0
dicttoxml==1.7.4
diff-match-patch==20200713
Django==3.0.2
django-cors-headers==3.7.0
django-crontab==0.7.1
django-redis==4.12.1
django-debug-toolbar==3.2
django-filter==2.4.0
django-import-export==2.3.0
django-model-utils==4.1.1
django-simpleui==2020.9.26
django-storage-qcloud==0.2.1
djangorestframework==3.11.1
et-xmlfile==1.0.1
idna==2.6
itsdangerous==1.1.0
jdcal==1.4.1
MarkupPy==1.14
mysqlclient==2.0.3
odfpy==1.4.1
openpyxl==3.0.6
pycryptodome==3.10.1
pytz==2021.1
PyYAML==5.4.1
qcloud-python-sts==3.0.5
redis==3.5.3
requests==2.25.1
simplejson==3.17.2
simpleui==4.0.2
six==1.15.0
sqlparse==0.4.1
tablib==3.0.0
urllib3==1.25.11
wechatpy==1.8.14
bs4==0.0.1
lxml==4.6.4
pandas==1.3.4
celery==4.4.7
importlib-metadata==4.8.3